Desktop/Cursor Apps/Permit Helper/backend/scrapers/permit_scraper.py
import requests
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from webdriver_manager.chrome import ChromeDriverManager
import time
import logging

class PermitScraper:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.sources = {
            "sf": self._scrape_sf_permits,
            "nyc": self._scrape_nyc_permits,
            "la": self._scrape_la_permits,
            # Add more sources as needed
        }
    
    def get_permits(self, address, city=None, state=None):
        """
        Get permits for a specific address
        
        Args:
            address (str): The address to search for
            city (str, optional): The city
            state (str, optional): The state
            
        Returns:
            list: List of permit dictionaries
        """
        # Determine which source to use based on city/state
        source = self._determine_source(city, state)
        
        if source in self.sources:
            return self.sources[source](address, city, state)
        else:
            # Default to a general search across multiple sources
            return self._general_permit_search(address, city, state)
    
    def _determine_source(self, city, state):
        """Determine which source to use based on city/state"""
        if not city or not state:
            return None
            
        city = city.lower()
        state = state.lower()
        
        if city == "san francisco" and state == "ca":
            return "sf"
        elif city == "new york" and state == "ny":
            return "nyc"
        elif city == "los angeles" and state == "ca":
            return "la"
        
        return None
    
    def _general_permit_search(self, address, city=None, state=None):
        """Search for permits across multiple sources"""
        # This is a placeholder for a more sophisticated search
        # In a real implementation, you might search multiple sources
        # or use a more general permit database
        
        return [
            {
                "id": "sample-1",
                "type": "Sample Permit",
                "status": "Pending",
                "issued_date": "2023-01-01",
                "description": "This is a sample permit for demonstration purposes",
                "address": address,
                "source": "Sample Source"
            }
        ]
    
    def _scrape_sf_permits(self, address, city=None, state=None):
        """Scrape permit data from San Francisco's permit database"""
        self.logger.info(f"Scraping SF permits for address: {address}")
        
        # This is a placeholder for the actual implementation
        # In a real implementation, you would:
        # 1. Set up Selenium or requests to access the SF permit website
        # 2. Navigate to the search page
        # 3. Enter the address
        # 4. Submit the form
        # 5. Parse the results
        
        # Placeholder data
        return [
            {
                "id": "sf-123456",
                "type": "Building Permit",
                "status": "Approved",
                "issued_date": "2023-02-15",
                "description": "Kitchen remodel",
                "address": address,
                "source": "San Francisco Planning Department"
            }
        ]
    
    def _scrape_nyc_permits(self, address, city=None, state=None):
        """Scrape permit data from NYC's permit database"""
        self.logger.info(f"Scraping NYC permits for address: {address}")
        
        # Placeholder data
        return [
            {
                "id": "nyc-789012",
                "type": "Construction Permit",
                "status": "In Review",
                "issued_date": "2023-03-10",
                "description": "New construction - residential",
                "address": address,
                "source": "NYC Department of Buildings"
            }
        ]
    
    def _scrape_la_permits(self, address, city=None, state=None):
        """Scrape permit data from LA's permit database"""
        self.logger.info(f"Scraping LA permits for address: {address}")
        
        # Placeholder data
        return [
            {
                "id": "la-345678",
                "type": "Electrical Permit",
                "status": "Completed",
                "issued_date": "2023-01-20",
                "description": "Electrical panel upgrade",
                "address": address,
                "source": "LA Department of Building and Safety"
            }
        ]
    
    def _setup_selenium(self):
        """Set up Selenium WebDriver"""
        chrome_options = Options()
        chrome_options.add_argument("--headless")
        chrome_options.add_argument("--no-sandbox")
        chrome_options.add_argument("--disable-dev-shm-usage")
        
        service = Service(ChromeDriverManager().install())
        driver = webdriver.Chrome(service=service, options=chrome_options)
        
        return driver 