Desktop/Cursor Apps/Permit Helper/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pymongo==4.6.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0 